refactor(workos): extract WorkOS protocol adapter from oauth.ts\n\nMove refreshWorkosToken and credentialsFromWorkos from oauth.ts to\nworkos.ts with injectable fetch for testability. workos.ts now owns\nall WorkOS protocol knowledge. oauth.ts becomes pure orchestration.
